MitMub/Back-end/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-appconf==1.0.5
django-cors-headers==3.5.0
django-imagekit==4.1.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
pilkit==2.0
Pillow==9.3.0
PyJWT==1.7.1
pytz==2020.4
six==1.16.0
sqlparse==0.4.1
